Keep the public tree to the Apache-2.0 SDK
Drop non-SDK documents and their CI wiring, fold the trademark note
into NOTICE, and keep release safety as a neutral
scripts/check_release.py (release tag matches pyproject version,
LICENSE+NOTICE in the dist, secrets/scratch denylist). Trim internal
references from docs, comments, the changelog, and the PR template.
