+++ NOTICE
@@ -6,4 +6,4 @@ Licensed under the Apache License, Version 2.0. See LICENSE.
 
 ContextDB and the ContextDB logo are trademarks of Atoms AI.
 The Apache License does not grant permission to use the ContextDB
-trademarks. See COMMERCIAL.md.
+trademarks.
