+++ contextdb/core/config.py
@@ -109,8 +109,8 @@ class ContextDBConfig(BaseSettings):
         description="Logging level (DEBUG, INFO, WARNING, ERROR, CRITICAL).",
     )
 
-    # Optional local pathways in the open SDK — not entitlements. Cloud
-    # features are never gated by a boolean in this package.
+    # Optional local pathways in the open SDK. No feature in this package
+    # is gated behind a hosted offering.
     enable_entity_graph: bool = Field(default=True)
     enable_multi_graph: bool = Field(default=False)
     enable_rl_manager: bool = Field(default=False)
