+++ pyproject.toml
@@ -128,9 +128,8 @@ addopts = "-v --tb=short"
 packages = ["contextdb"]
 
 # Defense in depth against accidentally shipping secrets or scratch files
-# in the sdist (the 0.1.0 release got burned by a stray .rtf containing a
-# token). Whatever lands here is excluded from both the sdist and the
-# wheel build contexts.
+# in the sdist. Whatever lands here is excluded from both the sdist and
+# the wheel build contexts.
 [tool.hatch.build]
 exclude = [
     "*.rtf",
@@ -146,13 +145,8 @@ exclude = [
     "*.pdf",
     "contextdb-*.md",
     "agentic-memory-*.md",
-    "contextdb_cloud/",
-    "cloud/",
     ".cursor/",
     "legal/signed/",
-    "scripts/setup-contextdb-cloud.sh",
-    "TASKS.md",
-    "PRD.md",
 ]
 
 [tool.hatch.build.targets.sdist]
@@ -170,11 +164,6 @@ exclude = [
     "*.pdf",
     "contextdb-*.md",
     "agentic-memory-*.md",
-    "contextdb_cloud/",
-    "cloud/",
     ".cursor/",
     "legal/signed/",
-    "scripts/setup-contextdb-cloud.sh",
-    "TASKS.md",
-    "PRD.md",
 ]
